fix(back): edit fields for correct type

+++ producer-service/pkg/models/item.go
@@ -1,7 +1,7 @@
 package models
 
 type Item struct {
-	ChrtID      int    `json:"chrt_id"`
+	ChrtID      int64    `json:"chrt_id"`
 	TrackNumber string `json:"track_number"`
 	Price       int    `json:"price"`
 	RID         string `json:"rid"`
@@ -9,7 +9,7 @@ type Item struct {
 	Sale        int    `json:"sale"`
 	Size        string `json:"size"`
 	TotalPrice  int    `json:"total_price"`
-	NmID        int    `json:"nm_id"`
+	NmID        int64    `json:"nm_id"`
 	Brand       string `json:"brand"`
 	Status      int    `json:"status"`
 }
